+++ src/main/java/myongari/backend/club/application/ClubFacade.java
@@ -3,9 +3,9 @@ package myongari.backend.club.application;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import myongari.backend.club.domain.Image;
-import myongari.backend.club.presentation.dto.ClubNamesAndDetail;
-import myongari.backend.club.presentation.dto.ClubSimple;
-import myongari.backend.club.presentation.dto.ClubSimplePage;
+import myongari.backend.club.dto.ClubNamesAndDetail;
+import myongari.backend.club.dto.ClubSimple;
+import myongari.backend.club.dto.ClubSimplePage;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
