+++ src/main/java/myongari/backend/club/application/ClubService.java
@@ -2,8 +2,8 @@ package myongari.backend.club.application;
 
 import lombok.RequiredArgsConstructor;
 import myongari.backend.club.application.port.ClubRepository;
-import myongari.backend.club.presentation.dto.ClubNamesAndDetail;
-import myongari.backend.club.presentation.dto.ClubSimple;
+import myongari.backend.club.dto.ClubNamesAndDetail;
+import myongari.backend.club.dto.ClubSimple;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
