+++ src/main/java/myongari/backend/club/infra/ClubDslRepository.java
@@ -10,9 +10,9 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import myongari.backend.club.domain.Club;
-import myongari.backend.club.presentation.dto.ClubName;
-import myongari.backend.club.presentation.dto.ClubNamesAndDetail;
-import myongari.backend.club.presentation.dto.ClubSimple;
+import myongari.backend.club.dto.ClubName;
+import myongari.backend.club.dto.ClubNamesAndDetail;
+import myongari.backend.club.dto.ClubSimple;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
